+++ src/components/documentBuilder/createNewElement.test.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import validUuidRegex from "@/vendors/validateUuid";
+import { isUUID } from "@/types/helpers";
 import { createNewElement } from "./createNewElement";
 import { DOCUMENT_ELEMENT_TYPES } from "./documentBuilderTypes";
 import { type BrickPipeline } from "@/bricks/types";
@@ -114,8 +114,8 @@ test("sets padding to zero for form", () => {
   expect((actual.config as any).pipeline.__value__[0].id).toBe(
     "@pixiebrix/form",
   );
-  expect((actual.config as any).pipeline.__value__[0].instanceId).toMatch(
-    validUuidRegex,
+  expect(isUUID((actual.config as any).pipeline.__value__[0].instanceId)).toBe(
+    true,
   );
   expect((actual.config as any).pipeline.__value__[0].config.className).toBe(
     "p-0",
