+++ src/components/fields/schemaFields/widgets/CodeEditorWidget.tsx
@@ -17,7 +17,7 @@
 
 import React, { Suspense } from "react";
 import { type SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
-import AceEditor from "@/vendors/AceEditor";
+import AceEditor from "@/components/AceEditor";
 import { useField } from "formik";
 
 const CodeEditorWidget: React.VFC<SchemaFieldProps> = ({ name }) => {
