+++ src/vendors/jQueryInitialize.js
